Remove the old type inference
